rename tests folder, add test, add coveragerc
