+++ .coveragerc
@@ -1,0 +1,7 @@
+[run]
+omit =
+    */site-packages/*
+    */dist-packages/*
+    */distutils/*
+    tests/*
+    patchmentation/utils/validator.py
